feat: Add real-time prominence detection with geometric mean fusion
- Add RT mode to syllable_detector with online calibration
- Add syllable_set_realtime_mode and syllable_recalibrate APIs
- Add DSP modules: AGC, MFCC, spectral flux, wavelet, high-freq energy
- Add web demo with Wasm integration
- Add Interspeech and Visceral Resonance paper drafts
- Add KissFFT external dependency

+++ src/dsp/zff.h
@@ -10,8 +10,10 @@ typedef struct {
   float trend_accum;
 } ZFF;
 
+#include <stddef.h>
+
 void zff_init(ZFF *z, int sample_rate, float trend_window_ms,
-              void *(*custom_alloc)(unsigned long long));
+              void *(*custom_alloc)(size_t));
 void zff_process(ZFF *z, float in, float *zff_out, float *slope_out);
 void zff_destroy(ZFF *z, void (*custom_free)(void *));
